+++ pvinspect/data/__init__.py
@@ -1,6 +1,6 @@
 '''Data loading and viewing'''
 
-from .io import save_images, read_module_image, read_module_images, \
+from .io import save_images, save_image, read_module_image, read_module_images, \
     read_partial_module_image, read_partial_module_images
 from .image import EL_IMAGE, PL_IMAGE
 from . import demo
@@ -8,5 +8,5 @@ from . import demo
 __all__ = [
         'read_module_image', 'read_module_images', 
         'read_partial_module_image', 'read_partial_module_images',
-        'save_images', 'demo', 'EL_IMAGE', 'PL_IMAGE'
+        'save_images', 'save_image', 'demo', 'EL_IMAGE', 'PL_IMAGE'
     ]