+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   * includes new reading methods `read_partial_module_image`and `read_partial_module_images`
 * Automatic clipping of outliers in viewer
   * includes `clip_low` and `clip_high` arguments in viewing methods
+* Add `save_image` to allow saving a single image
 
 ## 0.0.1
 * Initial version